Gui (#19)
* Started create_icon()

* Fixed fps drop when placing an object

* Fixed object being placed not following the camera

* Started getting ray traced icons and gui now hides automaticaly when unlocking camera

* Fixed object creation box and menu box not using BOX_ROUNDING_RADIUS for their corners rounding

* Finished object creation back-end

* Made slightly better icons, still needs work

* Prepared the ground for the adding of the ability to add lights

* Can now interact with cones

* Can now place and modify a light

* Can now change objects reflection

* Fixed write_centered_string_to_image()

* Started improving buttons look

* Fixed float input box outline

* Float input boxes now show a plus and a minus sign

* GUI now shows each float value of the selected object

* Updated GUI color

* Started modifying t_gui_box so that they don't require an image. Will result in lower memory usage, a slight performance increase on mac and a bigger one on linux

* Adapted all gui boxes to work with the new gui box image system

* Buttons now handle scroll

* Fixed light color not being set and updated correctly

* Updated scrolling behaviour on normal and position boxes

* Started init_object_list_box()

* Object list now functionnal

* Object list icons now update as the object changes

* Fixed object list sub boxes not being writen in the correct order

* Seperated object settings

* Slight refactoring

* Can now toggle checked pattern

* Can now change chekered pattern color

* Started working on the buttons to change checkered pattern size

* Fixed checkered pattern size buttons

* Moved checkerboard.c

* Fixed potential seg fault

* Sarted fixing checkerboard buttons

+++ src/gui/object_modification_box/init_color_picker/color_getter_functions.c
@@ -15,13 +15,13 @@
 int	get_increasing_color(float color, float x, float decal,
 				float ratio)
 {
-	return ((int)roundf(color * ((x - decal) / ratio)));
+	return (roundf(color * ((x - decal) / ratio)));
 }
 
 int	get_decreasing_color(float color, float x, float decal,
 				float ratio)
 {
-	return ((int)roundf(color * (1.0 - ((x - decal) / ratio))));
+	return (roundf(color * (1.0 - ((x - decal) / ratio))));
 }
 
 int	get_base_color(float color, float x, float decal, float ratio)
@@ -29,5 +29,5 @@ int	get_base_color(float color, float x, float decal, float ratio)
 	(void)x;
 	(void)decal;
 	(void)ratio;
-	return ((int)color);
+	return (color);
 }
